Ignore `compile_invalid_pyc_invalidation_mode` on all platforms
This test is failing most times for me when running nextest locally, so i'm deactivating it for now. I'm still not sure what the root cause here is. It seems to have something to do with python stdin not being ready immediately after we spawn the process and us being too fast.