Remove special casing from no solution error
In uv, the only pubgrub error that can occur is a `NoSolutionError`, and the only place it can occur is `unit_propagation`. By changing the return type on pubgrub's side (https://github.com/astral-sh/pubgrub/pull/28), we can remove the pattern matching and the `unreachable!()` calls.

Our pubgrub error used to have a two phased initialization, first mostly stubs in `solve[_tracked]()` and then adding the actual values in `resolve()`. When constructing the error in `solve` we already have all this context, so we can unify this to a regular constructor and remove the special casing in `resolve()` and `hints()`.