Fix `uv init .` (#5330)
## Summary

This PR avoids an `Invalid package name` error that occurs when using
`uv init .`. This is achieved by slightly reorganizing the code block to
determine the name after the path is canonicalized. The dot path is
expanded to the current directory, and the `file_name` then works as
expected.

Resolve #5329 .

---------

Co-authored-by: konstin <konstin@mailbox.org>