fix(venv): make relocatable activation scripts support ksh
It transpires that detecting the directory a script was sourced from
is non-trivial across `bash`, `ksh` and `zsh`.

The previous version was a one-liner and supported `bash` and `zsh` but
not `ksh`.

It is possible to keep the one-liner and add `ksh` support, but that
is mutually-exclusive with `zsh`.

Therefore, the only way to square this circle is to add an `if` block.
A silver lining here is that although longer, the script is probably
easier to follow as there is less code-golfing going on.