uv/tests: update snapshots from overlapping marker change
This commit is meant to group together test updates that are
"uninteresting." That is, most updates are just adding marker
expressions that are "always true" or establishing the forks in the
lock file.

I've tried to split out the interesting snapshot changes into subsequent
commits.

Some of the marker changes here are quite striking. I checked most of
them. Every one I checked, including the big ones, are all just fancy
ways of saying, "always true."